# perf: Avoid repeated traversal when creating scope